98percent finish

+++ public/js/adminSubjects.js
@@ -21,20 +21,46 @@ $(function(e) {
 
     $("table.list tbody").on("click", "button.btndel", function(e) {
         const current = $(this)
-
-        $.ajax({
-            type: "DELETE",
-            url: `/admin/deleteSubject?code=${current.attr("id")}`,
-            success: (res) => {
-                if(!res.operation) return alert("Something went wrong")
-
-                current.closest("tr").remove()
-
-                alert(res.msg)
+        
+        const swalWithBootstrapButtons = Swal.mixin({
+            customClass: {
+              confirmButton: "btn btn-success",
+              cancelButton: "btn btn-danger",
             },
-            error: (error) => {
-                console.log(error)
-                alert("Server Error")
+            buttonsStyling: true,
+          });
+
+          swalWithBootstrapButtons
+          .fire({
+            title: "Are you sure?",
+            text: "You won't be able to revert this!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonText: "Yes, delete it!",
+            cancelButtonText: "No, cancel!",
+            reverseButtons: true,
+          })
+        .then((result) => {
+            if (result.isConfirmed) {
+                $.ajax({
+                    type: "DELETE",
+                    url: `/admin/deleteSubject?code=${current.attr("id")}`,
+                    success: (res) => {
+                        if(!res.operation) return alert("Something went wrong")
+        
+                        current.closest("tr").remove()
+        
+                        swalWithBootstrapButtons.fire(
+                            "Deleted!",
+                            "Your file has been deleted.",
+                            "success"
+                          );
+                    },
+                    error: (error) => {
+                        console.log(error)
+                        alert("Server Error")
+                    }
+                })
             }
         })
     })
@@ -44,7 +70,7 @@ $(function(e) {
     $("#add-btn").click(function(e) {
 
         if(!subject_code.val() || !subject_name.val() || !subject_units.val() || !subject_year.val() || !subject_sem.val() || courses.val().length <= 0) return alert("Please complete all input")
-
+        
         $.ajax({
             type: "POST",
             url: "/admin/jeffersonpogi",
@@ -82,36 +108,36 @@ $(function(e) {
             }
         })
 
-        $.ajax({
-            type: "POST",
-            url: "/student/subjectCourse",
-            headers: {
-                "Content-type": "application/json"
-            },
-            data: JSON.stringify({
-                code: subject_code.val(),
-                s_name: subject_name.val(),
-                units: subject_units.val(),
-            }),
-            success: (res) => {
-                if (!res.operation) return alert(res.msg)
+        // $.ajax({
+        //     type: "POST",
+        //     url: "/student/subjectCourse",
+        //     headers: {
+        //         "Content-type": "application/json"
+        //     },
+        //     data: JSON.stringify({
+        //         code: subject_code.val(),
+        //         s_name: subject_name.val(),
+        //         units: subject_units.val(),
+        //     }),
+        //     success: (res) => {
+        //         if (!res.operation) return alert(res.msg)
 
-                alert(res.msg)
+        //         alert(res.msg)
 
-                $("table.list tbody").append(`
-                <tr>
-                    <td>${subject_code.val()}</td>
-                    <td>${subject_name.val()}</td>
-                    <td>${subject_units.val()}</td>
-                    <td>${courses.val()}</td>
-                </tr>
-                `)
-            },
-            error: (error) => {
-                alert("Something went wrong")
-                console.log(error)
-            }
-        })
+        //         $("table.list tbody").append(`
+        //         <tr>
+        //             <td>${subject_code.val()}</td>
+        //             <td>${subject_name.val()}</td>
+        //             <td>${subject_units.val()}</td>
+        //             <td>${courses.val()}</td>
+        //         </tr>
+        //         `)
+        //     },
+        //     error: (error) => {
+        //         alert("Something went hi")
+        //         console.log(error)
+        //     }
+        // })
     })
 
 })
